LocalMessage: override getContentType() and make sure headers are loaded
This makes sure that MessageExtractor is able to extract the
correct charset from a LocalMessage (see issue #604).

There was a problem with the possible exceptions thrown, however:
Loading headers results in a MessagingException being thrown, but
so far Part.getContentType() did not specify this.

This means there were two choices:
1) Have Part.getContentType() specify 'throws MessagingException'
   at the top level.
2) Don't let loadHeaders() throw a MessagingException at all and
   remove all the throws declaration on the depending methods.

Choice 1 turned out to be hard because it had significant influence
to other parts of the code.

Choice 2 is easy but probably not a good solution.

This commit uses choice 2.